Marked abstract types abstract

+++ lib/IR/List.h
@@ -16,9 +16,15 @@ template <typename T, typename Parent>
 using NodeWithParent = utl::ilist_node_with_parent<T, Parent>;
 
 template <typename T>
-struct DynAllocator: std::allocator<T> {
+class DynAllocator: public std::allocator<T> {
+public:
     void destroy(T* ptr) {
-        visit(*ptr, [](auto& obj) { std::destroy_at(&obj); });
+        if constexpr (requires(T * ptr) { ptr->privateDestroy(); }) {
+            ptr->privateDestroy();
+        }
+        else {
+            visit(*ptr, [](auto& obj) { std::destroy_at(&obj); });
+        }
     }
 
     void deallocate(T* ptr, size_t count) { operator delete(ptr, count); }
